backend/online_teacher/requirements.txt
asgiref
crispy-bootstrap5
distlib
dj-database-url
Django
django-crispy-forms
django-environ
filelock
more-itertools
pillow
platformdirs
psycopg2-binary
sqlparse
typing_extensions
virtualenv
django-email
django-environ
django-filter
djangorestframework
django-cors-headers
